split out shard creation into module
